Use a single step and input for ABI/solc output #3196
